+++ Runtime/Core/PropertyGraphNode.cs
@@ -5,6 +5,7 @@ namespace skysepehru.Core.PropertyRegistry
     internal class PropertyGraphNode
     {
         public PropertyFilter PropertyFilter;
+        public PropertyFilter[] InputFilters;
         public List<PropertyGraphNode> Inputs;
         public List<PropertyGraphNode> Outputs;
         public IPropertyCalculator Calculator;